Move vanilla markup functions back to MinkContext
