+++ src/Drupal/DrupalExtension/Context/MarkupContext.php
@@ -2,7 +2,6 @@
 
 namespace Drupal\DrupalExtension\Context;
 
-use Behat\Mink\Exception\UnsupportedDriverActionException;
 use Behat\MinkExtension\Context\RawMinkContext;
 
 /**
@@ -29,95 +28,6 @@ class MarkupContext extends RawMinkContext {
     }
 
     return $regionObj;
-  }
-
-  /**
-   * @Then I should see the link :link
-   */
-  public function assertLinkVisible($link) {
-    $element = $this->getSession()->getPage();
-    $result = $element->findLink($link);
-
-    try {
-      if ($result && !$result->isVisible()) {
-        throw new \Exception(sprintf("No link to '%s' on the page %s", $link, $this->getSession()->getCurrentUrl()));
-      }
-    }
-    catch (UnsupportedDriverActionException $e) {
-      // We catch the UnsupportedDriverActionException exception in case
-      // this step is not being performed by a driver that supports javascript.
-      // All other exceptions are valid.
-    }
-
-    if (empty($result)) {
-      throw new \Exception(sprintf("No link to '%s' on the page %s", $link, $this->getSession()->getCurrentUrl()));
-    }
-  }
-
-  /**
-   * @Then I should not see the link :link
-   */
-  public function assertNotLinkVisible($link) {
-    $element = $this->getSession()->getPage();
-    $result = $element->findLink($link);
-
-    try {
-      if ($result && $result->isVisible()) {
-        throw new \Exception(sprintf("The link '%s' was present on the page %s and was not supposed to be", $link, $this->getSession()->getCurrentUrl()));
-      }
-    }
-    catch (UnsupportedDriverActionException $e) {
-      // We catch the UnsupportedDriverActionException exception in case
-      // this step is not being performed by a driver that supports javascript.
-      // All other exceptions are valid.
-    }
-
-    if ($result) {
-      throw new \Exception(sprintf("The link '%s' was present on the page %s and was not supposed to be", $link, $this->getSession()->getCurrentUrl()));
-    }
-  }
-
-  /**
-   * @Then I (should )see the heading :heading
-   */
-  public function assertHeading($heading) {
-    $element = $this->getSession()->getPage();
-    foreach (array('h1', 'h2', 'h3', 'h4', 'h5', 'h6') as $tag) {
-      $results = $element->findAll('css', $tag);
-      foreach ($results as $result) {
-        if ($result->getText() == $heading) {
-          return;
-        }
-      }
-    }
-    throw new \Exception(sprintf("The text '%s' was not found in any heading on the page %s", $heading, $this->getSession()->getCurrentUrl()));
-  }
-
-  /**
-   * @Then I (should )not see the heading :heading
-   */
-  public function assertNotHeading($heading) {
-    $element = $this->getSession()->getPage();
-    foreach (array('h1', 'h2', 'h3', 'h4', 'h5', 'h6') as $tag) {
-      $results = $element->findAll('css', $tag);
-      foreach ($results as $result) {
-        if ($result->getText() == $heading) {
-          throw new \Exception(sprintf("The text '%s' was found in a heading on the page %s", $heading, $this->getSession()->getCurrentUrl()));
-        }
-      }
-    }
-  }
-
-  /**
-   * @Then I (should ) see the button :button
-   * @Then I (should ) see the :button button
-   */
-  public function assertButton($button) {
-    $element = $this->getSession()->getPage();
-    $buttonObj = $element->findButton($button);
-    if (empty($buttonObj)) {
-      throw new \Exception(sprintf("The button '%s' was not found on the page %s", $button, $this->getSession()->getCurrentUrl()));
-    }
   }
 
   /**
